src/requirements.txt
pandas==1.2.3
PyDriller==2.0
jira==3.0.1
GitPython==3.1.24
python_dateutil==2.8.2
